refactor: Application / GameController 분리
