+++ src/main/java/racingcar/Application.java
@@ -1,0 +1,10 @@
+package racingcar;
+
+import racingcar.controller.GameController;
+
+public class Application {
+
+    public static void main(String[] args) {
+        new GameController().play();
+    }
+}
